fix(developer): enchancer for adding is_dev_account to tracking events
